backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.84
botocore==1.29.84
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
openai==0.27.0
pydantic==1.10.5
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
